requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
python-multipart==0.0.9

SQLAlchemy==2.0.29
alembic==1.13.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

httpx==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1

redis==5.0.4
celery==5.3.6

python-magic==0.4.27
Pillow==10.2.0

structlog==24.1.0

pytest==8.1.1
pytest-asyncio==0.23.6
black==24.3.0
flake8==7.0.0